+++ Module.AudioSourcesComparer/GUI/AbstractViewModels/IVkAudioVM.cs
@@ -4,7 +4,7 @@ namespace Module.AudioSourcesComparer.GUI.AbstractViewModels
 {
     public interface IVkAudioVM
     {
-        int Id { get; } // todo mb long
+        long Id { get; } // todo mb long
         string Artist { get; }
         string Title { get; }
